requirements.txt
click==8.0.3
flask==2.0.2
itsdangerous==2.0.1
keras==2.7.0
matplotlib==3.3.4
mtcnn==0.1.1
numpy==1.21.4
Pillow==8.4.0
scikit-image==0.19.0
scikit-learn==0.24.1
sklearn==0.0
tensorflow==2.7.0
werkzeug==2.0.2